requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
requests==2.32.5
python-dotenv==1.2.1
supabase==2.27.2
langchain==1.2.8
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-google-genai==4.2.1
langchain-groq==1.1.2
google-genai==1.65.0
google-generativeai==0.8.5
pydantic==2.11.9
